Add port configuration for heroku
Heroku needs the service to obey the PORT environment variable. This
does that. It also introduces spock testing and the all-open plugin so
that spock can mock kotlin classes.

The OpenForTesting annotation is similar to what I've used
VisibleForTesting in the past except that it is actually hooked up to a
complier plugin rather than just a visual indicator for the reader.