A little service cleanup and logging (#11)
This commit adds signal handling so that when dev mode stops it returns
a valid return code rather than null. That was kinda bugging me. I added
a logback.xml so that I'm not getting spammed with debug logging when I
only really care about info level logging.

Also this is now logging the environment variables being passed in. This
will give me a cheat sheet for when I try to figure out what heroku is
sending my service.

I added a test for the new logging, which allowed me to explore
how to do named bindings in kotlin.

Finally, I added a logger for the github webhook endpoint I'm planning
on developing.