first pass at adding react (#16)
This work adds a basic create react app to the project. It also removes the server side page rendering in favor of this new app. I've added builds that push the UI to s3. Finally I made the UI serve up dir configurable so that it will work proper in java dev mode.