House keeping (#37)
The readme was blank so I added some ramblings about the project. I update the ui code so that it is able to
run, and I updated the version of gradle to 6.8. I changed the default port to 8081 because my development machine
has a persistent service running on 8080. I should probably do something cooler like detect that 8080 is in use
and choose the next open port, but I'm too lazy for that, this is my day off damn it.

Finally I've added the concept of user.gradle so that I can start working with the adobe api.